[Refactor] Prepare FragmentContext primitive deps for ExecRuntime
