+++ be/src/exec/pipeline/fragment_context.h
@@ -44,7 +44,6 @@
 #include "gen_cpp/PlanNodes_types.h"
 #include "gen_cpp/QueryPlanExtra_types.h"
 #include "gen_cpp/Types_types.h"
-#include "runtime/runtime_filter_worker.h"
 #include "runtime/runtime_state.h"
 #include "storage/primitive/predicate_tree_params.h"
 
@@ -52,6 +51,7 @@ namespace starrocks {
 
 class FragmentAttachment;
 class FragmentDictState;
+class RuntimeFilterPort;
 
 namespace pipeline {
